Client 3 4 refactor (#77)
* Refactor rendering system

Currently rendering is statically polymorphised with use of CRTP
and std::variant/std::visit. Rendering and view changing happens
as a call to current view. Input handling is also carried out by
the view with the same interface forced on all of them. This allowed
to get rid of template mess in game_gui and decouple state a bit.
Unfortunately still 3 forward declarations of client context were
needed in rendering files because input handling requires views to
be able to modify the `view` field of the context. Thus they need to
be aware of their own existence in another structure. A better solution
would be to return possible new state and check if its nullptr or
whatever.

* Add networking object

* Add next networking commit from another branch

* Separate files into appropriate folders and change singleton

* clang-format